--- bench/bench-cl3a.ros
@@ -1,37 +1,0 @@
-#!/bin/sh
-#|-*- mode:lisp -*-|#
-#|
-exec ros -Q -- $0 "$@"
-|#
-
-
-#+sbcl (sb-posix:chdir "../")
-#+ccl (ccl:cwd "../")
-(ql:quickload "cl3a" :silent t)
-#+sbcl (sb-posix:chdir "./bench")
-#+ccl (ccl:cwd "./bench")
-
-
-(defun run-bench (n m)
-  (declare (type fixnum n m))
-  (let ((va (make-array (list n)
-                        :element-type 'double-float))
-        (vb (make-array (list n)
-                        :element-type 'double-float)))
-    (declare (type (simple-array double-float (*)) va vb))
-    (dotimes (i n)
-      (setf (aref va i) (random 1d0))
-      (setf (aref vb i) (random 1d0)))
-    (time (loop :repeat m :do (cl3a:dv*v va vb)))))
-
-
-(defun main (n-str m-str &rest argv)
-  (declare (ignorable argv))
-  (let ((n (parse-integer n-str))
-        (m (parse-integer m-str)))
-    (declare (type fixnum n m))
-    ; #+sbcl (sb-profile:profile "CL3A" "CL3A.UTILITIES" "CL3A.DOTPROD")
-    (run-bench n m)
-    ; #+sbcl (sb-profile:report)
-    ; #+sbcl (sb-profile:unprofile)
-    ))
